# public/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PRISM API",
    "version": "1.0.0",
    "description": "API surface for PRISM. Includes test status from the latest backend smoke test."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    },
    {
      "url": "https://prism-jade-nine.vercel.app",
      "description": "Production"
    }
  ],
  "x-test-report": {
    "reportUrl": "/docs/test-report.json"
  },
  "paths": {
    "/api/auth/providers": {
      "get": {
        "summary": "List NextAuth providers",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/auth/csrf": {
      "get": {
        "summary": "Get CSRF token",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/auth/callback/credentials": {
      "post": {
        "summary": "Credentials login",
        "responses": {
          "200": { "description": "OK" },
          "302": { "description": "Redirect" }
        }
      }
    },
    "/api/auth/session": {
      "get": {
        "summary": "Get session",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/session/validate": {
      "post": {
        "summary": "Validate session token",
        "responses": {
          "200": { "description": "OK" }
        }
      }
    },
    "/api/session/check-conflict": {
      "post": {
        "summary": "Check session conflict",
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Missing userId" }
        }
      }
    },
    "/api/session/sessions": {
      "get": {
        "summary": "List active sessions",
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Logout all sessions",
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Invalid request" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/trips": {
      "get": {
        "summary": "List trips",
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create trip",
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/trips/{id}": {
      "get": {
        "summary": "Get trip",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Invalid trip ID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      },
      "put": {
        "summary": "Update trip",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Invalid trip ID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "summary": "Delete trip",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "OK" },
          "400": { "description": "Invalid trip ID" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Not found" }
        }
      }
    }
  }
}